Refactoring tests
* luatest updated to 0.4.0 version
* tests splitted by groups, and to integration and unit
* added before_suite, before_each, after_each hooks
* added tests shuffling
* deleted test_split_uri test, because it tested builtin 'uri' rock
